Models/requirements.txt
# Core framework
fastapi>=0.104.0
uvicorn[standard]>=0.22.0

# ML & Data
numpy>=1.24.0
tensorflow-intel==2.17.0          # locks protobuf<5.0.0
protobuf>=3.20.3,<5.0.0
scikit-learn==1.4.1.post1
scipy>=1.10.0
pandas>=2.0.0
joblib>=1.3.1

# Imaging & I/O
Pillow>=9.0.0
python-multipart>=0.0.6

# Web scraping & HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Environment & Configuration
python-dotenv>=1.0.0

# Geocoding & Location
geocoder>=1.38.1
geopy>=2.3.0

# Weather API
openmeteo-requests>=1.1.0
requests-cache>=1.0.0
retry-requests>=2.0.0

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-google-genai>=0.0.6
langchain-chroma>=0.1.0

# Vector store dependencies
chromadb>=0.4.0

# Search tools dependencies
duckduckgo-search>=3.9.0
youtube-search-python>=1.6.0

# Additional utilities
requests>=2.31.0